+++ src/honeyhive/_generated/services/Sessions_service.py
@@ -13,11 +13,7 @@ def getSession(
 
     base_path = api_config.base_path
     path = f"/v1/sessions/{session_id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -50,11 +46,7 @@ def deleteSession(
 
     base_path = api_config.base_path
     path = f"/v1/sessions/{session_id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {