+++ src/honeyhive/_generated/services/async_Events_service.py
@@ -21,11 +21,7 @@ async def getEvents(
 
     base_path = api_config.base_path
     path = f"/events"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {
         "dateRange": dateRange,
         "filters": filters,
@@ -68,11 +64,7 @@ async def createEvent(
 
     base_path = api_config.base_path
     path = f"/events"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -108,11 +100,7 @@ async def updateEvent(
 
     base_path = api_config.base_path
     path = f"/events"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -157,11 +145,7 @@ async def getEventsChart(
 
     base_path = api_config.base_path
     path = f"/v1/events/chart"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {
         "dateRange": dateRange,
         "filters": filters,
@@ -207,11 +191,7 @@ async def getEventsBySessionId(
 
     base_path = api_config.base_path
     path = f"/v1/events/{id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -250,11 +230,7 @@ async def deleteEvent(
 
     base_path = api_config.base_path
     path = f"/v1/events/{id}"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -289,11 +265,7 @@ async def exportEvents(
 
     base_path = api_config.base_path
     path = f"/v1/events/export"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -333,11 +305,7 @@ async def createModelEvent(
 
     base_path = api_config.base_path
     path = f"/events/model"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -373,11 +341,7 @@ async def createEventBatch(
 
     base_path = api_config.base_path
     path = f"/events/batch"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -415,11 +379,7 @@ async def createModelEventBatch(
 
     base_path = api_config.base_path
     path = f"/events/model/batch"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {