+++ src/honeyhive/_generated/services/async_Metrics_service.py
@@ -16,11 +16,7 @@ async def getMetrics(
 
     base_path = api_config.base_path
     path = f"/v1/metrics"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {"type": type, "id": id}
 
     query_params = {
@@ -55,11 +51,7 @@ async def createMetric(
 
     base_path = api_config.base_path
     path = f"/v1/metrics"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -95,11 +87,7 @@ async def updateMetric(
 
     base_path = api_config.base_path
     path = f"/v1/metrics"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {
@@ -135,11 +123,7 @@ async def deleteMetric(
 
     base_path = api_config.base_path
     path = f"/v1/metrics"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {"metric_id": metric_id}
 
     query_params = {
@@ -174,11 +158,7 @@ async def runMetric(
 
     base_path = api_config.base_path
     path = f"/v1/metrics/run_metric"
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": f"Bearer { api_config.get_access_token() }",
-    }
+    headers = api_config.get_default_headers()
     query_params: Dict[str, Any] = {}
 
     query_params = {